+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changelog
 ### Changed
 - Changed `dukt\social\github\loginproviders\Github::getOauthProvider()` method’s visibility to `public`.
 - Removed `dukt\social\github\loginproviders\Github::getProfile()`.
+- Renamed `dukt\social\github\loginproviders\Github::getDefaultScope()` to `dukt\social\github\loginproviders\Github::getDefaultOauthScope()`.
 
 ## 3.0.2 - 2018-05-18