+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog
 - Removed `dukt\social\github\loginproviders\Github::getProfile()`.
 - Renamed `dukt\social\github\loginproviders\Github::getDefaultScope()` to `dukt\social\github\loginproviders\Github::getDefaultOauthScope()`.
 - Updated `dukt\social\github\loginproviders\Github::getOauthProvider()` to take into account the new OAuth provider config introduced in Social 2.0.0-beta.10.
+- Updated `dukt/social` composer dependency to `^2.0.0-beta.10`.
 
 ## 3.0.2 - 2018-05-18